Improved simulation mode in GUI
Optimized pre-parsing to stop when all necessary MetaData are found.
Fixed spurious error message when running a simulation
Improved progress dialog and added a "Pause/Resume" function.

+++ CNC/gcode.h
@@ -3,4 +3,4 @@ void showDistanceInfo( );
 tStatus preParse( char* cmd );
 tStatus doGcode( char* cmd );
 void setSimfile( FILE* file );
-void getTheoricalPos(float* x, float* y, float* z);
+void getTheoricalPos(double* x, double* y, double* z);
